Seperating the logic more | Partially broken
