+++ media_recommendation_engine/src/utils/mod.rs
@@ -20,6 +20,9 @@ pub mod templates;
 mod settings;
 pub use settings::ServerSettings;
 
+mod streaming;
+pub use streaming::{Session, StreamingSessions};
+
 macro_rules! relative {
     ($path:expr) => {
         if cfg!(windows) {
